requirements.txt
fastapi~=0.115.0
uvicorn~=0.31.0
sqlalchemy~=2.0.35
asyncpg
passlib~=1.7.4
bcrypt
python-jose
python-dotenv~=1.0.1
pytest
pytest-asyncio
httpx
pyjwt~=2.9.0
requests~=2.32.3
config~=0.5.1
pydantic~=2.9.2
python-magic~=0.4.27
pillow~=11.0.0
locust~=2.32.0